Fix type safety warnings: SetPort casts, switch comparisons, pointer comparisons
Fixed 21 warnings by adding proper type casts:
- WindowPtr → GrafPtr casts for SetPort (9 warnings eliminated)
- OSType cast for alias kind switch statement (6 warnings eliminated)
- Handle → short casts for window proc comparisons (6 warnings eliminated)

Files modified:
- src/Finder/folder_window.c: Added GrafPtr cast to SetPort
- src/ControlPanels/cdev_desktop.c: Added GrafPtr casts to 4 SetPort calls
- src/ControlManager/ControlManagerCore.c: Added GrafPtr casts to 2 SetPort calls
- src/TextEdit/TextEditTest.c: Added GrafPtr casts to 2 SetPort calls
- src/Finder/alias_manager.c: Cast aliasKind to OSType for switch statement
- src/WindowManager/WindowManagerHelpers.c: Cast windowDefProc to short for resource ID comparisons

Warnings: 744 → 723 (-21)

+++ src/WindowManager/WindowManagerHelpers.c
@@ -199,10 +199,10 @@ short WM_GetWindowLayer(WindowPtr window) {
         return 2;  /* Floating window */
     }
 
-    /* Check if modal dialog */
-    if (window->windowDefProc == dBoxProc ||
-        window->windowDefProc == plainDBox ||
-        window->windowDefProc == altDBoxProc) {
+    /* Check if modal dialog - cast windowDefProc to short for resource ID comparison */
+    if ((short)(long)window->windowDefProc == dBoxProc ||
+        (short)(long)window->windowDefProc == plainDBox ||
+        (short)(long)window->windowDefProc == altDBoxProc) {
         return 1;  /* Modal dialog */
     }
 
@@ -222,9 +222,9 @@ Boolean WM_IsFloatingWindow(WindowPtr window) {
 
 Boolean WM_IsAlertDialog(WindowPtr window) {
     if (!window) return false;
-    return (window->windowDefProc == dBoxProc ||
-            window->windowDefProc == plainDBox ||
-            window->windowDefProc == altDBoxProc);
+    return ((short)(long)window->windowDefProc == dBoxProc ||
+            (short)(long)window->windowDefProc == plainDBox ||
+            (short)(long)window->windowDefProc == altDBoxProc);
 }
 
 /* Window overlap testing */
